explode list-entry module
